+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.times;
+package roomescape.repository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
