+++ src/test/java/roomescape/controller/IsolateWithJdbcTemplateTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservations;
+package roomescape.controller;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
